companies/novartis/pyproject.toml
[project]
name = "novartis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pytest-playwright>=0.7.2",
    "sqlite-minutils>=4.0.3",
]
